+++ src/place_global/density_legalizer.hpp
@@ -73,14 +73,14 @@ class DensityLegalizer : public HierarchicalDensityPlacement {
    * @brief Update the x coordinates to target
    */
   void updateCellTargetX(const std::vector<float> &cellTargetX) {
-    assert(cellTargetX.size() == nbCells());
+    assert((int) cellTargetX.size() == nbCells());
     cellTargetX_ = cellTargetX;
   }
   /**
    * @brief Update the y coordinates to target
    */
   void updateCellTargetY(const std::vector<float> &cellTargetY) {
-    assert(cellTargetY.size() == nbCells());
+    assert((int) cellTargetY.size() == nbCells());
     cellTargetY_ = cellTargetY;
   }
