+++ src/place_global/transportation.cpp
@@ -174,10 +174,10 @@ void TransportationProblem::costsFromIntegers(
 }
 
 void TransportationProblem::check() const {
-  if (demands_.size() != nbSources()) {
+  if ((int) demands_.size() != nbSources()) {
     throw std::runtime_error("Inconsistant demands");
   }
-  if (capacities_.size() != nbSinks()) {
+  if ((int) capacities_.size() != nbSinks()) {
     throw std::runtime_error("Inconsistant capacities");
   }
   for (DemandType c : demands_) {
@@ -191,23 +191,23 @@ void TransportationProblem::check() const {
     }
   }
 
-  if (costs_.size() != nbSinks()) {
+  if ((int) costs_.size() != nbSinks()) {
     throw std::runtime_error(
         "Inconsistant cost size (first dimension is sinks)");
   }
   for (const auto& c : costs_) {
-    if (c.size() != nbSources()) {
+    if ((int) c.size() != nbSources()) {
       throw std::runtime_error(
           "Inconsistant cost size (second dimension is sources)");
     }
   }
 
-  if (allocations_.size() != nbSinks()) {
+  if ((int) allocations_.size() != nbSinks()) {
     throw std::runtime_error(
         "Inconsistant allocation size (first dimension is sinks)");
   }
   for (const auto& c : allocations_) {
-    if (c.size() != nbSources()) {
+    if ((int) c.size() != nbSources()) {
       throw std::runtime_error(
           "Inconsistant allocation size (second dimension is sources)");
     }
@@ -375,7 +375,7 @@ void TransportationProblem::setAllocations(
 
 void TransportationProblem::setAssignment(const std::vector<int>& assignment) {
   allocations_.assign(nbSinks(), std::vector<DemandType>(nbSources(), 0LL));
-  if (assignment.size() > nbSources()) {
+  if ((int) assignment.size() > nbSources()) {
     throw std::runtime_error(
         "Assignment should be no larger than the number of sources");
   }
